requirements.txt
python-dotenv
starlette
tortoise-orm[asyncpg]
uvicorn